adding changes from grid_mods
