+++ plugins/grid-bookmark/src/GridBookmarkWidget/components/AssemblySelector.tsx
@@ -1,48 +1,38 @@
 import React from 'react'
 import { observer } from 'mobx-react'
 
-import {
-  Select,
-  MenuItem,
-  FormControl,
-  InputLabel,
-  Checkbox,
-  ListItemText,
-  OutlinedInput,
-  ListItemIcon,
-  SelectChangeEvent,
-} from '@mui/material'
-import { makeStyles } from 'tss-react/mui'
+import { Checkbox, MenuItem, SelectChangeEvent, TextField } from '@mui/material'
 
 // locals
 import { GridBookmarkModel } from '../model'
+import { makeStyles } from 'tss-react/mui'
+import { getSession } from '@jbrowse/core/util'
 
-const useStyles = makeStyles()(() => ({
-  bold: {
-    fontWeight: 700,
+const useStyles = makeStyles()({
+  textfield: {
+    padding: 5,
   },
-}))
+})
 
-const AssemblySelector = observer(function ({
+const GridBookmarkAssemblySelector = observer(function AssemblySelector({
   model,
 }: {
   model: GridBookmarkModel
 }) {
+  const { assemblyManager } = getSession(model)
   const { classes } = useStyles()
   const { validAssemblies, selectedAssemblies, setSelectedAssemblies } = model
-  const noAssemblies = validAssemblies.length === 0 ? true : false
-  const label = 'Select assemblies'
-  const isAllSelected =
-    validAssemblies.length > 0 &&
-    selectedAssemblies.length === validAssemblies.length
+  const allSelected =
+    validAssemblies.size > 0 &&
+    selectedAssemblies.length === validAssemblies.size
 
   const handleChange = (event: SelectChangeEvent<string[]>) => {
     const value = event.target.value
-    if (value.at(-1) === 'all') {
+    if (value.at(-1) === 'All') {
       setSelectedAssemblies(
-        selectedAssemblies.length === validAssemblies.length
+        selectedAssemblies.length === validAssemblies.size
           ? []
-          : validAssemblies,
+          : [...validAssemblies],
       )
       return
     }
@@ -50,41 +40,33 @@ const AssemblySelector = observer(function ({
   }
 
   return (
-    <FormControl disabled={noAssemblies}>
-      <InputLabel id="select-assemblies-label">{label}</InputLabel>
-      <Select
-        labelId="select-assemblies-label"
-        id="select-assemblies"
-        multiple
-        value={selectedAssemblies}
-        onChange={handleChange}
-        input={<OutlinedInput label={label} />}
-        renderValue={selected => selected.join(', ')}
-      >
-        <MenuItem value="all">
-          <ListItemIcon>
-            <Checkbox
-              checked={isAllSelected}
-              indeterminate={
-                selectedAssemblies.length > 0 &&
-                selectedAssemblies.length < validAssemblies.length
-              }
-            />
-          </ListItemIcon>
-          <ListItemText
-            classes={{ primary: classes.bold }}
-            primary="Select All"
-          />
+    <TextField
+      select
+      variant="outlined"
+      fullWidth
+      className={classes.textfield}
+      SelectProps={{
+        multiple: true,
+        value: selectedAssemblies,
+        // @ts-ignore
+        onChange: handleChange,
+        // @ts-ignore
+        renderValue: selected => selected.join(', '),
+      }}
+      label="Select assembly"
+    >
+      <MenuItem value="All">
+        <Checkbox checked={allSelected} />
+        All
+      </MenuItem>
+      {[...validAssemblies].map(name => (
+        <MenuItem key={name} value={name}>
+          <Checkbox checked={selectedAssemblies.includes(name)} />
+          {assemblyManager.get(name)?.displayName || name}
         </MenuItem>
-        {validAssemblies.map(name => (
-          <MenuItem key={name} value={name}>
-            <Checkbox checked={selectedAssemblies.includes(name)} />
-            <ListItemText primary={name} />
-          </MenuItem>
-        ))}
-      </Select>
-    </FormControl>
+      ))}
+    </TextField>
   )
 })
 
-export default AssemblySelector
+export default GridBookmarkAssemblySelector
