+++ plugins/grid-bookmark/src/GridBookmarkWidget/components/GridBookmarkWidget.tsx
@@ -15,15 +15,11 @@ import { GridBookmarkModel } from '../model'
 
 const useStyles = makeStyles()({
   card: {
-    display: 'flex',
-    flexFlow: 'column',
-    margin: '5px',
-    padding: '5px',
-    gap: '5px',
+    marginTop: 5,
   },
 })
 
-const GridBookmarkWidget = observer(function ({
+const GridBookmarkWidget = observer(function GridBookmarkWidget({
   model,
 }: {
   model: GridBookmarkModel
@@ -43,8 +39,9 @@ const GridBookmarkWidget = observer(function ({
         <DeleteBookmarks model={model} />
       </div>
       <Alert severity="info">
-        Click or double click the <strong>label</strong> field to notate your
-        bookmark.
+        Click and type within the <strong>label</strong> field to annotate your
+        bookmark. Double click the <strong>label</strong> field to do so within
+        a dialog.
       </Alert>
       <AssemblySelector model={model} />
       <BookmarkGrid model={model} />
